+++ Scripts/Data/PlayerData.cs
@@ -3,12 +3,12 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "ScriptableObjects/Player", order = 1)]
 public class PlayerData : ScriptableObject
 {
-    public int startHp;
+    public float startHp;
     public float startSpeed;
-    public int startManaShield;
-    public int startArmor;
-    public int startHpRegen;
-    public int startManaShieldRegen;
+    public float startManaShield;
+    public float startArmor;
+    public float startHpRegen;
+    public float startManaShieldRegen;
     public float timeToStartHpRegenAfterTakeDamage;
     public float timeToStartManaShieldRegenAfterTakeDamage;
 }
